+++ lib/vector-api-client/src/lib.rs
@@ -15,6 +15,7 @@ mod client;
 /// GraphQL queries
 pub mod gql;
 mod subscription;
+pub mod test;
 
 pub use client::*;
 pub use subscription::*;
