+++ lib/vector-api-client/tests/queries/component_links.graphql
